# Conflicts:
#	.gitignore
#	broker/model/block.go
#	broker/model/delegation_reward_message.go
#	broker/model/mint_params.go
#	broker/model/multi_send_message.go
#	broker/model/proposal.go
#	broker/model/proposal_vote_message.go
#	broker/model/send_message.go
#	broker/model/staking_params.go
#	broker/model/validator.go

+++ broker/model/staking_params.go
@@ -5,9 +5,9 @@ import "time"
 type (
 	SParams struct {
 		UnbondingTime     time.Duration `json:"unbonding_time"`
-		MaxValidators     uint32        `json:"max_validators"`
-		MaxEntries        uint32        `json:"max_entries"`
-		HistoricalEntries uint32        `json:"historical_entries"`
+		MaxValidators     uint64        `json:"max_validators"`
+		MaxEntries        uint64        `json:"max_entries"`
+		HistoricalEntries uint64        `json:"historical_entries"`
 		BondDenom         string        `json:"bond_denom"`
 		MinCommissionRate float64       `json:"min_commission_rate"`
 	}
